Working resampler has been implemented, finally! Since this moment, is no more need to resample musics and sounds into 44100 hz! So, Vanilia tracks are will work fine.

+++ _Libs/SDL2_mixer_modified/ADLMIDI/adlmidi.h
@@ -100,7 +100,7 @@ extern void adl_reset(struct ADL_MIDIPlayer*device);
 extern void adl_close(struct ADL_MIDIPlayer *device);
 
 /*Take a sample buffer*/
-extern int  adl_play(struct ADL_MIDIPlayer*device, int sampleCount, int out []);
+extern int  adl_play(struct ADL_MIDIPlayer*device, int sampleCount, short out[]);
 
 #ifdef __cplusplus
      }
